Service Worker and ServiceWorkerPlugin refactor. (#1832)
* Refactored sw.js

* Modified webpack to use the new service worker folder structure.

* Refactored ServiceWorkerPlugin.

* Refactored configureWebpack.js

* Fix fileDependencies check.

* Minor.

* Adding png to image loader test regex.

* Enabling service worker debugging.

* for of syntax change because jest does not understand if const or let is not mentioned. Weird.

* Tests update.

* Code refactor to fix build issues.

* Renaming default variable names.

* Update packages/pwa-buildpack/lib/Utilities/getServiceWorkerConfig.js


* Created MAX_NUM_OF_IMAGES_TO_CACHE default.

* Added JSDocs for ServiceWorkerPlugin.

* ServiceWorkerPlugin tests refactor.
