+++ packages/pwa-buildpack/lib/WebpackTools/PWADevServer.js
@@ -169,38 +169,46 @@ const PWADevServer = {
                 let middleware;
                 const gatheringQueryTabs = new Promise((resolve, reject) => {
                     const { compiler } = server.middleware.context;
-                    compiler.hooks.done.tap('PWADevServer', async stats => {
-                        const queryFilePaths = [];
-                        for (const filename of stats.compilation
-                            .fileDependencies) {
-                            if (filename.endsWith('.graphql')) {
-                                queryFilePaths.push(filename);
+                    compiler.hooks.done.tap(
+                        'PWADevServer',
+                        async ({ stats }) => {
+                            /**
+                             * Stats in an array because we have 2 webpack child
+                             * compilations, 1 for client and other for service worker.
+                             */
+                            const queryFilePaths = [];
+                            for (const { compilation } of stats) {
+                                for (const filename of compilation.fileDependencies) {
+                                    if (filename.endsWith('.graphql')) {
+                                        queryFilePaths.push(filename);
+                                    }
+                                }
+                            }
+                            try {
+                                resolve(
+                                    await Promise.all(
+                                        queryFilePaths.map(async queryFile => {
+                                            const query = await readFile(
+                                                queryFile,
+                                                'utf8'
+                                            );
+                                            const name = path.relative(
+                                                context,
+                                                queryFile
+                                            );
+                                            return {
+                                                endpoint,
+                                                name,
+                                                query
+                                            };
+                                        })
+                                    )
+                                );
+                            } catch (e) {
+                                reject(e);
                             }
                         }
-                        try {
-                            resolve(
-                                await Promise.all(
-                                    queryFilePaths.map(async queryFile => {
-                                        const query = await readFile(
-                                            queryFile,
-                                            'utf8'
-                                        );
-                                        const name = path.relative(
-                                            context,
-                                            queryFile
-                                        );
-                                        return {
-                                            endpoint,
-                                            name,
-                                            query
-                                        };
-                                    })
-                                )
-                            );
-                        } catch (e) {
-                            reject(e);
-                        }
-                    });
+                    );
                 });
                 /* istanbul ignore next: dummy next() function not testable */
                 const noop = () => {};
