+++ packages/pwa-buildpack/lib/WebpackTools/__tests__/configureWebpack.spec.js
@@ -67,7 +67,7 @@ test('produces a webpack config and friendly manifest plugin', async () => {
         .statsAsDirectory()
         .statsAsFile()
         .productionEnvironment();
-    const config = await configureWebpack({ context: '.' });
+    const { clientConfig: config } = await configureWebpack({ context: '.' });
     expect(config).toMatchObject({
         context: '.',
         mode: 'production',
@@ -115,9 +115,12 @@ test('works in developer mode from cli', async () => {
         .statsAsDirectory()
         .statsAsMissing()
         .productionEnvironment();
-    await expect(
-        configureWebpack({ context: '.', env: { mode: 'development' } })
-    ).resolves.toHaveProperty('mode', 'development');
+    const { clientConfig } = await configureWebpack({
+        context: '.',
+        env: { mode: 'development' }
+    });
+
+    expect(clientConfig).toHaveProperty('mode', 'development');
 });
 
 test('works in developer mode from fallback', async () => {
@@ -125,10 +128,9 @@ test('works in developer mode from fallback', async () => {
         .statsAsDirectory()
         .statsAsMissing()
         .devEnvironment();
-    await expect(configureWebpack({ context: '.' })).resolves.toHaveProperty(
-        'mode',
-        'development'
-    );
+    const { clientConfig } = await configureWebpack({ context: '.' });
+
+    expect(clientConfig).toHaveProperty('mode', 'development');
 });
 
 test('errors when mode unrecognized', async () => {
@@ -147,7 +149,7 @@ test('handles special flags', async () => {
         .statsAsFile()
         .productionEnvironment();
 
-    const config = await configureWebpack({
+    const { clientConfig } = await configureWebpack({
         context: '.',
         vendor: ['jest'],
         special: {
@@ -168,7 +170,7 @@ test('handles special flags', async () => {
         }
     });
     expect(
-        config.module.rules.find(({ use }) =>
+        clientConfig.module.rules.find(({ use }) =>
             use.some(({ loader }) => /^graphql\-tag/.test(loader))
         ).include
     ).toHaveLength(3);
