+++ packages/venia-concept/src/ServiceWorker/Utilities/imageCacheHandler.js
@@ -1,48 +1,16 @@
-const thirtyDays = 30 * 24 * 60 * 60;
-workbox.core.skipWaiting();
-workbox.core.clientsClaim();
-
-workbox.routing.registerRoute(
-    '/',
-    new workbox.strategies.StaleWhileRevalidate()
-);
-
-workbox.routing.registerRoute(
-    new RegExp('\\.html$'),
-    new workbox.strategies.NetworkFirst()
-);
-
-workbox.routing.registerRoute(
-    new RegExp('/.\\.js$'),
-    new workbox.strategies.StaleWhileRevalidate()
-);
-
-const catalogCacheName = 'catalog';
+import { CATALOG_CACHE_NAME } from '../defaults';
 
 const getWidth = url => Number(new URLSearchParams(url.search).get('width'));
 
 const isCatalogImage = ({ url }) => url.pathname.startsWith('/media/catalog');
 
-const isResizedCatalogImage = ({ url }) =>
+export const isResizedCatalogImage = ({ url }) =>
     isCatalogImage({ url }) && !isNaN(getWidth(url));
 
-const catalogCacheHandler = new workbox.strategies.StaleWhileRevalidate({
-    cacheName: catalogCacheName,
-    plugins: [
-        new workbox.expiration.Plugin({
-            maxEntries: 60,
-            maxAgeSeconds: thirtyDays
-        }),
-        new workbox.cacheableResponse.Plugin({
-            statuses: [0, 200]
-        })
-    ]
-});
-
-const findSameOrLargerImage = async (url, request) => {
+export const findSameOrLargerImage = async (url, request) => {
     const requestedWidth = getWidth(url);
 
-    const cache = await caches.open(catalogCacheName);
+    const cache = await caches.open(CATALOG_CACHE_NAME);
     const cachedSources = await cache.matchAll(request, {
         ignoreSearch: true
     });
@@ -92,35 +60,3 @@ const findSameOrLargerImage = async (url, request) => {
         return best.candidate;
     }
 };
-
-workbox.routing.registerRoute(
-    isResizedCatalogImage,
-    ({ url, request, event }) => {
-        const sameOrLargerImagePromise = findSameOrLargerImage(url, request);
-        event.waitUntil(sameOrLargerImagePromise);
-        return sameOrLargerImagePromise.then(
-            response =>
-                response || catalogCacheHandler.handle({ request, event })
-        );
-    }
-);
-
-workbox.routing.registerRoute(
-    /\.(?:png|gif|jpg|jpeg|svg)$/,
-    new workbox.strategies.CacheFirst({
-        cacheName: 'images',
-        plugins: [
-            new workbox.expiration.Plugin({
-                maxEntries: 60,
-                maxAgeSeconds: thirtyDays // 30 Days
-            })
-        ]
-    })
-);
-
-workbox.precaching.precacheAndRoute(self.__precacheManifest || []);
-
-// This "catch" handler is triggered when any of the other routes fail to
-// generate a response.
-
-// TODO: Add fallbacks
