+++ packages/venia-concept/src/ServiceWorker/defaults.js
@@ -1,0 +1,5 @@
+export const THIRTY_DAYS = 30 * 24 * 60 * 60;
+
+export const CATALOG_CACHE_NAME = 'catalog';
+
+export const MAX_NUM_OF_IMAGES_TO_CACHE = 60;
