+++ packages/venia-concept/src/ServiceWorker/setupWorkbox.js
@@ -1,0 +1,11 @@
+export default function() {
+    importScripts(
+        'https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js'
+    );
+
+    workbox.core.skipWaiting();
+
+    workbox.core.clientsClaim();
+
+    workbox.precaching.precacheAndRoute(self.__precacheManifest || []);
+}
