+++ packages/venia-concept/src/ServiceWorker/sw.js
@@ -1,0 +1,6 @@
+import setupWorkbox from './setupWorkbox';
+import registerRoutes from './registerRoutes';
+
+setupWorkbox();
+
+registerRoutes();
