+++ packages/venia-concept/webpack.config.js
@@ -7,11 +7,12 @@ const HTMLWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = async env => {
     const mediaUrl = await getMediaURL();
+
     global.MAGENTO_MEDIA_BACKEND_URL = mediaUrl;
 
     const unionAndInterfaceTypes = await getUnionAndInterfaceTypes();
 
-    const config = await configureWebpack({
+    const { clientConfig, serviceWorkerConfig } = await configureWebpack({
         context: __dirname,
         vendor: [
             'apollo-cache-inmemory',
@@ -56,9 +57,9 @@ module.exports = async env => {
      * supports the `module.noParse` option in Webpack, documented here:
      * https://webpack.js.org/configuration/module/#modulenoparse
      */
-    config.module.noParse = [/braintree\-web\-drop\-in/];
-    config.plugins = [
-        ...config.plugins,
+    clientConfig.module.noParse = [/braintree\-web\-drop\-in/];
+    clientConfig.plugins = [
+        ...clientConfig.plugins,
         new DefinePlugin({
             /**
              * Make sure to add the same constants to
@@ -77,5 +78,5 @@ module.exports = async env => {
         })
     ];
 
-    return config;
+    return [clientConfig, serviceWorkerConfig];
 };
